pyproject.toml
[project]
name = "endscopetool"
version = "0.1.0"
description = "Python implementation of the endscopetool Android application"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pillow",
    "opencv",
    "trio"
]

[project.scripts]
endscope = "endscopetool:cli_main"

[tool.setuptools]
py-modules = ["endscopetool", "fake_endscope", "transports"]
